[#60] fix: 모델 마이그레이션을 위한 모델 재정의
