+++ TapTap/Projects/Core/Sources/SwiftDataModel/CategoryItem.swift
@@ -20,42 +20,4 @@ public struct CategoryIcon: Codable, Hashable {
   }
 }
 
-@Model
-public final class CategoryItem: Identifiable, Codable, Equatable, Sendable {
-  public var id: UUID = UUID()
-  public var categoryName: String = "" // 카테고리 이름
-  public var createdAt: Date = Date()
-  public var icon: CategoryIcon = CategoryIcon()
-  
-  @Relationship(inverse: \ArticleItem.category) public var links: [ArticleItem]? = []
-  
-  public init(
-    categoryName: String,
-    icon: CategoryIcon
-  ) {
-    self.id = UUID()
-    self.categoryName = categoryName
-    self.createdAt = Date()
-    self.icon = icon
-  }
-  
-  enum CodingKeys: String, CodingKey {
-    case id, categoryName, createdAt, icon
-  }
-  
-  public required init(from decoder: Decoder) throws {
-    let container = try decoder.container(keyedBy: CodingKeys.self)
-    self.id = try container.decode(UUID.self, forKey: .id)
-    self.categoryName = try container.decode(String.self, forKey: .categoryName)
-    self.createdAt = try container.decode(Date.self, forKey: .createdAt)
-    self.icon = try container.decode(CategoryIcon.self, forKey: .icon)
-  }
-  
-  public func encode(to encoder: Encoder) throws {
-    var container = encoder.container(keyedBy: CodingKeys.self)
-    try container.encode(id, forKey: .id)
-    try container.encode(categoryName, forKey: .categoryName)
-    try container.encode(createdAt, forKey: .createdAt)
-    try container.encode(icon, forKey: .icon)
-  }
-}
+// CategoryItem은 TapTapSchema.swift에서 VersionedSchema로 정의되어 있습니다.
